+++ gavicore/src/gavicore/ui/field/context.py
@@ -75,12 +75,14 @@ class FieldContext:
         return LayoutManager(layout_function, views).layout(self)
 
     def create_property_fields(self) -> dict[str, Field]:
+        assert isinstance(self.meta.properties, dict)
         return {
             prop_name: self.create_child_field(prop_meta)
             for prop_name, prop_meta in self.meta.properties.items()
         }
 
     def create_item_field(self) -> Field:
+        assert isinstance(self.meta.items, FieldMeta)
         return self.create_child_field(self.meta.items)
 
     def create_child_field(self, child_meta: FieldMeta) -> Field: