+++ CHANGES.md
@@ -2,6 +2,10 @@
 
 ### Enhancements
 
+- TODO: Describe the new UI generator and its use
+  - Added `gavicore.ui` and removed `cuiman.gui.component` (TODO) 
+  - Skipped support for `additionalParameters` in process descriptions;
+  - Skipped support for experimental `ClientConfig.is_advanced_input()` 
 - The Cuiman client package has been enhanced by _job result openers_,
   which ease working with the results of a process job (#65):
     - Client classes now have a method